+++ service/medialive/api.go
@@ -4,6 +4,8 @@ package medialive
 
 import (
 	"fmt"
+	"io"
+	"time"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/awsutil"
@@ -1634,6 +1636,96 @@ func (c *MediaLive) DescribeInputDevice(input *DescribeInputDeviceInput) (*Descr
 // for more information on using Contexts.
 func (c *MediaLive) DescribeInputDeviceWithContext(ctx aws.Context, input *DescribeInputDeviceInput, opts ...request.Option) (*DescribeInputDeviceOutput, error) {
 	req, out := c.DescribeInputDeviceRequest(input)
+	req.SetContext(ctx)
+	req.ApplyOptions(opts...)
+	return out, req.Send()
+}
+
+const opDescribeInputDeviceThumbnail = "DescribeInputDeviceThumbnail"
+
+// DescribeInputDeviceThumbnailRequest generates a "aws/request.Request" representing the
+// client's request for the DescribeInputDeviceThumbnail operation. The "output" return
+// value will be populated with the request's response once the request completes
+// successfully.
+//
+// Use "Send" method on the returned Request to send the API call to the service.
+// the "output" return value is not valid until after Send returns without error.
+//
+// See DescribeInputDeviceThumbnail for more information on using the DescribeInputDeviceThumbnail
+// API call, and error handling.
+//
+// This method is useful when you want to inject custom logic or configuration
+// into the SDK's request lifecycle. Such as custom headers, or retry logic.
+//
+//
+//    // Example sending a request using the DescribeInputDeviceThumbnailRequest method.
+//    req, resp := client.DescribeInputDeviceThumbnailRequest(params)
+//
+//    err := req.Send()
+//    if err == nil { // resp is now filled
+//        fmt.Println(resp)
+//    }
+//
+// See also, https://docs.aws.amazon.com/goto/WebAPI/medialive-2017-10-14/DescribeInputDeviceThumbnail
+func (c *MediaLive) DescribeInputDeviceThumbnailRequest(input *DescribeInputDeviceThumbnailInput) (req *request.Request, output *DescribeInputDeviceThumbnailOutput) {
+	op := &request.Operation{
+		Name:       opDescribeInputDeviceThumbnail,
+		HTTPMethod: "GET",
+		HTTPPath:   "/prod/inputDevices/{inputDeviceId}/thumbnailData",
+	}
+
+	if input == nil {
+		input = &DescribeInputDeviceThumbnailInput{}
+	}
+
+	output = &DescribeInputDeviceThumbnailOutput{}
+	req = c.newRequest(op, input, output)
+	return
+}
+
+// DescribeInputDeviceThumbnail API operation for AWS Elemental MediaLive.
+//
+// Get the latest thumbnail data for the input device.
+//
+// Returns awserr.Error for service API and SDK errors. Use runtime type assertions
+// with awserr.Error's Code and Message methods to get detailed information about
+// the error.
+//
+// See the AWS API reference guide for AWS Elemental MediaLive's
+// API operation DescribeInputDeviceThumbnail for usage and error information.
+//
+// Returned Error Types:
+//   * BadRequestException
+//
+//   * InternalServerErrorException
+//
+//   * ForbiddenException
+//
+//   * BadGatewayException
+//
+//   * NotFoundException
+//
+//   * GatewayTimeoutException
+//
+//   * TooManyRequestsException
+//
+// See also, https://docs.aws.amazon.com/goto/WebAPI/medialive-2017-10-14/DescribeInputDeviceThumbnail
+func (c *MediaLive) DescribeInputDeviceThumbnail(input *DescribeInputDeviceThumbnailInput) (*DescribeInputDeviceThumbnailOutput, error) {
+	req, out := c.DescribeInputDeviceThumbnailRequest(input)
+	return out, req.Send()
+}
+
+// DescribeInputDeviceThumbnailWithContext is the same as DescribeInputDeviceThumbnail with the addition of
+// the ability to pass a context and additional request options.
+//
+// See DescribeInputDeviceThumbnail for details on how to use this API operation.
+//
+// The context must be non-nil and will be used for request cancellation. If
+// the context is nil a panic will occur. In the future the SDK may create
+// sub-contexts for http.Requests. See https://golang.org/pkg/context/
+// for more information on using Contexts.
+func (c *MediaLive) DescribeInputDeviceThumbnailWithContext(ctx aws.Context, input *DescribeInputDeviceThumbnailInput, opts ...request.Option) (*DescribeInputDeviceThumbnailOutput, error) {
+	req, out := c.DescribeInputDeviceThumbnailRequest(input)
 	req.SetContext(ctx)
 	req.ApplyOptions(opts...)
 	return out, req.Send()
@@ -5246,9 +5338,10 @@ type AudioDescription struct {
 	// Audio codec settings.
 	CodecSettings *AudioCodecSettings `locationName:"codecSettings" type:"structure"`
 
-	// Indicates the language of the audio output track. Only used if languageControlMode
-	// is useConfigured, or there is no ISO 639 language code specified in the input.
-	LanguageCode *string `locationName:"languageCode" min:"3" type:"string"`
+	// RFC 5646 language code representing the language of the audio output track.
+	// Only used if languageControlMode is useConfigured, or there is no ISO 639
+	// language code specified in the input.
+	LanguageCode *string `locationName:"languageCode" min:"1" type:"string"`
 
 	// Choosing followInput will cause the ISO 639 language code of the output to
 	// follow the ISO 639 language code of the input. The languageCode will be used
@@ -5288,8 +5381,8 @@ func (s *AudioDescription) Validate() error {
 	if s.AudioSelectorName == nil {
 		invalidParams.Add(request.NewErrParamRequired("AudioSelectorName"))
 	}
-	if s.LanguageCode != nil && len(*s.LanguageCode) < 3 {
-		invalidParams.Add(request.NewErrParamMinLen("LanguageCode", 3))
+	if s.LanguageCode != nil && len(*s.LanguageCode) < 1 {
+		invalidParams.Add(request.NewErrParamMinLen("LanguageCode", 1))
 	}
 	if s.Name == nil {
 		invalidParams.Add(request.NewErrParamRequired("Name"))
@@ -9307,6 +9400,113 @@ func (s *DescribeInputDeviceOutput) SetType(v string) *DescribeInputDeviceOutput
 	return s
 }
 
+type DescribeInputDeviceThumbnailInput struct {
+	_ struct{} `type:"structure"`
+
+	// Accept Header
+	//
+	// Accept is a required field
+	Accept *string `location:"header" locationName:"accept" type:"string" required:"true" enum:"AcceptHeader"`
+
+	// InputDeviceId is a required field
+	InputDeviceId *string `location:"uri" locationName:"inputDeviceId" type:"string" required:"true"`
+}
+
+// String returns the string representation
+func (s DescribeInputDeviceThumbnailInput) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s DescribeInputDeviceThumbnailInput) GoString() string {
+	return s.String()
+}
+
+// Validate inspects the fields of the type to determine if they are valid.
+func (s *DescribeInputDeviceThumbnailInput) Validate() error {
+	invalidParams := request.ErrInvalidParams{Context: "DescribeInputDeviceThumbnailInput"}
+	if s.Accept == nil {
+		invalidParams.Add(request.NewErrParamRequired("Accept"))
+	}
+	if s.InputDeviceId == nil {
+		invalidParams.Add(request.NewErrParamRequired("InputDeviceId"))
+	}
+	if s.InputDeviceId != nil && len(*s.InputDeviceId) < 1 {
+		invalidParams.Add(request.NewErrParamMinLen("InputDeviceId", 1))
+	}
+
+	if invalidParams.Len() > 0 {
+		return invalidParams
+	}
+	return nil
+}
+
+// SetAccept sets the Accept field's value.
+func (s *DescribeInputDeviceThumbnailInput) SetAccept(v string) *DescribeInputDeviceThumbnailInput {
+	s.Accept = &v
+	return s
+}
+
+// SetInputDeviceId sets the InputDeviceId field's value.
+func (s *DescribeInputDeviceThumbnailInput) SetInputDeviceId(v string) *DescribeInputDeviceThumbnailInput {
+	s.InputDeviceId = &v
+	return s
+}
+
+type DescribeInputDeviceThumbnailOutput struct {
+	_ struct{} `type:"structure" payload:"Body"`
+
+	Body io.ReadCloser `locationName:"body" type:"blob"`
+
+	ContentLength *int64 `location:"header" locationName:"Content-Length" type:"long"`
+
+	ContentType *string `location:"header" locationName:"Content-Type" type:"string" enum:"ContentType"`
+
+	ETag *string `location:"header" locationName:"ETag" type:"string"`
+
+	LastModified *time.Time `location:"header" locationName:"Last-Modified" type:"timestamp"`
+}
+
+// String returns the string representation
+func (s DescribeInputDeviceThumbnailOutput) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s DescribeInputDeviceThumbnailOutput) GoString() string {
+	return s.String()
+}
+
+// SetBody sets the Body field's value.
+func (s *DescribeInputDeviceThumbnailOutput) SetBody(v io.ReadCloser) *DescribeInputDeviceThumbnailOutput {
+	s.Body = v
+	return s
+}
+
+// SetContentLength sets the ContentLength field's value.
+func (s *DescribeInputDeviceThumbnailOutput) SetContentLength(v int64) *DescribeInputDeviceThumbnailOutput {
+	s.ContentLength = &v
+	return s
+}
+
+// SetContentType sets the ContentType field's value.
+func (s *DescribeInputDeviceThumbnailOutput) SetContentType(v string) *DescribeInputDeviceThumbnailOutput {
+	s.ContentType = &v
+	return s
+}
+
+// SetETag sets the ETag field's value.
+func (s *DescribeInputDeviceThumbnailOutput) SetETag(v string) *DescribeInputDeviceThumbnailOutput {
+	s.ETag = &v
+	return s
+}
+
+// SetLastModified sets the LastModified field's value.
+func (s *DescribeInputDeviceThumbnailOutput) SetLastModified(v time.Time) *DescribeInputDeviceThumbnailOutput {
+	s.LastModified = &v
+	return s
+}
+
 type DescribeInputInput struct {
 	_ struct{} `type:"structure"`
 
@@ -11998,7 +12198,7 @@ type H264Settings struct {
 	ParDenominator *int64 `locationName:"parDenominator" min:"1" type:"integer"`
 
 	// Pixel Aspect Ratio numerator.
-	ParNumerator *int64 `locationName:"parNumerator" type:"integer"`
+	ParNumerator *int64 `locationName:"parNumerator" min:"1" type:"integer"`
 
 	// H.264 Profile.
 	Profile *string `locationName:"profile" type:"string" enum:"H264Profile"`
@@ -12100,6 +12300,9 @@ func (s *H264Settings) Validate() error {
 	}
 	if s.ParDenominator != nil && *s.ParDenominator < 1 {
 		invalidParams.Add(request.NewErrParamMinValue("ParDenominator", 1))
+	}
+	if s.ParNumerator != nil && *s.ParNumerator < 1 {
+		invalidParams.Add(request.NewErrParamMinValue("ParNumerator", 1))
 	}
 	if s.QvbrQualityLevel != nil && *s.QvbrQualityLevel < 1 {
 		invalidParams.Add(request.NewErrParamMinValue("QvbrQualityLevel", 1))
@@ -12518,7 +12721,10 @@ type H265Settings struct {
 	// when it is constrained by themaximum bitrate. Recommended if you or your
 	// viewers pay for bandwidth.CBR: Quality varies, depending on the video complexity.
 	// Recommended only if you distributeyour assets to devices that cannot handle
-	// variable bitrates.
+	// variable bitrates.Multiplex: This rate control mode is only supported (and
+	// is required) when the video is beingdelivered to a MediaLive Multiplex in
+	// which case the rate control configuration is controlledby the properties
+	// within the Multiplex Program.
 	RateControlMode *string `locationName:"rateControlMode" type:"string" enum:"H265RateControlMode"`
 
 	// Sets the scan type of the output to progressive or top-field-first interlaced.
@@ -14861,9 +15067,7 @@ type InputPrepareScheduleActionSettings struct {
 	// The name of the input attachment that should be prepared by this action.
 	// If no name is provided, the action will stop the most recent prepare (if
 	// any) when activated.
-	//
-	// InputAttachmentNameReference is a required field
-	InputAttachmentNameReference *string `locationName:"inputAttachmentNameReference" type:"string" required:"true"`
+	InputAttachmentNameReference *string `locationName:"inputAttachmentNameReference" type:"string"`
 
 	// Settings to let you create a clip of the file input, in order to set up the
 	// input to ingest only a portion of the file.
@@ -14889,9 +15093,6 @@ func (s InputPrepareScheduleActionSettings) GoString() string {
 // Validate inspects the fields of the type to determine if they are valid.
 func (s *InputPrepareScheduleActionSettings) Validate() error {
 	invalidParams := request.ErrInvalidParams{Context: "InputPrepareScheduleActionSettings"}
-	if s.InputAttachmentNameReference == nil {
-		invalidParams.Add(request.NewErrParamRequired("InputAttachmentNameReference"))
-	}
 	if s.InputClippingSettings != nil {
 		if err := s.InputClippingSettings.Validate(); err != nil {
 			invalidParams.AddNested("InputClippingSettings", err.(request.ErrInvalidParams))
@@ -23817,6 +24018,12 @@ const (
 	Ac3MetadataControlUseConfigured = "USE_CONFIGURED"
 )
 
+// Accept Header
+const (
+	// AcceptHeaderImageJpeg is a AcceptHeader enum value
+	AcceptHeaderImageJpeg = "image/jpeg"
+)
+
 // Afd Signaling
 const (
 	// AfdSignalingAuto is a AfdSignaling enum value
@@ -24076,6 +24283,11 @@ const (
 
 	// ChannelStateUpdateFailed is a ChannelState enum value
 	ChannelStateUpdateFailed = "UPDATE_FAILED"
+)
+
+const (
+	// ContentTypeImageJpeg is a ContentType enum value
+	ContentTypeImageJpeg = "image/jpeg"
 )
 
 // The status of the action to synchronize the device configuration. If you